Set OPENBLAS_NUM_THREADS to 1
Make sure that OpenBLAS is restricted to 1 thread. As parallelism is
handled by having multiple Python processes performing computations on
different chunks, there is no need for this sort of parallelism to clash
with lower level parallelism like OpenBLASes. After all the Python
process parallelism allows the greatest degree of parallelism for the
longest period of time whereas OpenBLAS parallelism is basically
transient as it only exists for very specific operations in only some
steps.