Use `jakirkham/centos_conda` base image
Instead of using `jakirkham/centos_drmaa_conda`, which contains SGE and
DRMAA, use the non-SGE CentOS 6 image, `jakirkham/centos_conda`. This
makes it a bit more portable as the image doesn't require `sudo` in the
`ENTRYPOINT`. Also should make it a better fit for using with
Singularity on the cluster. Update the entrypoint scripts run for
testing as well (given this change).

Note: The SGE-based `nanshe` image will still be maintained under the
`sge` branch. Just it will no longer represent `latest`.