Unpin `tifffile`
This reverts commit 68e820e68f763a5c8826cdcf92094c362f11f250.

Previously we needed to pin `tifffile` as version 0.13+ introduced
breaking changes, which were not compatible with `nanshe` or
`dask-imread` (the latter used in the workflow). As such, we pinned
`tifffile` to a working version so that `nanshe` could behave as
expected. However `tifffile` has made some changes that fix
`dask-imread` and `nanshe` has also made changes to be more
accommodating towards `tifffile`'s new behavior. As such, we can now
unpin `tifffile` and use the latest version.