Skip moving around `nanshe` package
As `conda` 4.4+ seems to be confused as to where the `nanshe` package is
coming from and removing the `nanshe` package temporarily for testing
has little utility as far as testing is concerned (since both the
package and the downloaded source use the same code either way), simply
keep the installed package around during testing. Should avoid these
issues with `conda` recognizing `nanshe` was in the local index. Also
should avoid some unnecessary noise in the Docker layers.